chore, refactor clean up code
